+++ src/main.ts
@@ -95,7 +95,8 @@ class Irrigation extends utils.Adapter {
         await this.createRateLimitStates();
 
         this.valves = this.config2.valves.map(
-            (valveConfig, index) => new ValveController(this, index, valveConfig, this.rateLimiter),
+            (valveConfig, index) =>
+                new ValveController(this, index, valveConfig, this.rateLimiter, () => this.valves),
         );
         for (const valve of this.valves) {
             await valve.init();